Serve large content in smaller chunks
If we try to server a huge file (a DVD image) then gio will not like to
send a buffer size that's more that's in gigabytes.

Sending smaller chunks seems to work, for now we're sending 5Megs. Maybe
it's possible to create bigger buffer sizes as sending a 400Megs file
doesn't cause problems.